fix: numpy.float64 crash in json.dumps - ROOT CAUSE of zero signals
indicators.rsi() returned numpy.float64 instead of Python float.
json.dumps(metadata) in insert_signal crashed with TypeError but was
hidden by logger.debug. Signals were never inserted into DB.

Fixes:
- indicators.py: rsi() now returns float() explicitly
- strategist: all metadata values wrapped in float() for safety
- strategist: confidence min 0.3 instead of 0.0 (meaningless)
- strategist: logger.debug → logger.error (no more hidden crashes)

+++ python_agents/indicators.py
@@ -26,7 +26,7 @@ def rsi(prices: np.ndarray, period: int = 14) -> Optional[float]:
     if avg_loss == 0:
         return 100.0
     rs = avg_gain / avg_loss
-    return 100.0 - (100.0 / (1.0 + rs))
+    return float(100.0 - (100.0 / (1.0 + rs)))
 
 
 def macd(prices: np.ndarray, fast: int = 12, slow: int = 26,
